GK_07_LU_efficient
This code includes a number of improvements.  Variables such as 'Rounds_played' have been initialised at the start of the main routine AFTER the functions.  The  list containing the tokens has been defined outside of the While loop as it does not need to be defined each time a round is played.

The decoration around the feedback statements is not achieved by using a function (to avoid repeated code) and incrementing the number of rounds played occurs toward the start of the while loop (rather than being coded separately as part of each 'if / elif / else' statement.

From the users' perspective, one of the biggest changes is having the introduction separate from the error message that occurs if users enter invalid data when asked how much they would like to spend.  This game could be further improved by asking users if they have played the game before - if they say 'no', then instructions could be displayed, if they say 'yes', then the game could be launched without the instructions.